requirements.txt
streamlit
pandas
random
matplotlib.pyplot
numpy
io
